+++ lib/src/direction.dart
@@ -3,39 +3,39 @@ library piecemeal.src.direction;
 import 'vec.dart';
 
 class Direction extends VecBase implements Vec {
-  static const NONE = const Direction(0, 0);
-  static const N    = const Direction(0, -1);
-  static const NE   = const Direction(1, -1);
-  static const E    = const Direction(1, 0);
-  static const SE   = const Direction(1, 1);
-  static const S    = const Direction(0, 1);
-  static const SW   = const Direction(-1, 1);
-  static const W    = const Direction(-1, 0);
-  static const NW   = const Direction(-1, -1);
+  static const none = const Direction(0, 0);
+  static const n    = const Direction(0, -1);
+  static const ne   = const Direction(1, -1);
+  static const e    = const Direction(1, 0);
+  static const se   = const Direction(1, 1);
+  static const s    = const Direction(0, 1);
+  static const sw   = const Direction(-1, 1);
+  static const w    = const Direction(-1, 0);
+  static const nw   = const Direction(-1, -1);
 
   /// The eight cardinal and intercardinal directions.
-  static const ALL = const [N, NE, E, SE, S, SW, W, NW];
+  static const all = const [n, ne, e, se, s, sw, w, nw];
 
   /// The four cardinal directions: north, south, east, and west.
-  static const CARDINAL = const [N, E, S, W];
+  static const cardinal = const [n, e, s, w];
 
   /// The four directions between the cardinal ones: northwest, northeast,
   /// southwest and southeast.
-  static const INTERCARDINAL = const [NE, SE, SW, NW];
+  static const intercardinal = const [ne, se, sw, nw];
 
   const Direction(int x, int y) : super(x, y);
 
   Direction get rotateLeft45 {
     switch (this) {
-      case NONE: return NONE;
-      case N: return NW;
-      case NE: return N;
-      case E: return NE;
-      case SE: return E;
-      case S: return SE;
-      case SW: return S;
-      case W: return SW;
-      case NW: return W;
+      case none: return none;
+      case n: return nw;
+      case ne: return n;
+      case e: return ne;
+      case se: return e;
+      case s: return se;
+      case sw: return s;
+      case w: return sw;
+      case nw: return w;
     }
 
     throw "unreachable";
@@ -43,15 +43,15 @@ class Direction extends VecBase implements Vec {
 
   Direction get rotateRight45 {
     switch (this) {
-      case NONE: return NONE;
-      case N: return NE;
-      case NE: return E;
-      case E: return SE;
-      case SE: return S;
-      case S: return SW;
-      case SW: return W;
-      case W: return NW;
-      case NW: return N;
+      case none: return none;
+      case n: return ne;
+      case ne: return e;
+      case e: return se;
+      case se: return s;
+      case s: return sw;
+      case sw: return w;
+      case w: return nw;
+      case nw: return n;
     }
 
     throw "unreachable";
@@ -59,15 +59,15 @@ class Direction extends VecBase implements Vec {
 
   Direction get rotateLeft90 {
     switch (this) {
-      case NONE: return NONE;
-      case N: return W;
-      case NE: return NW;
-      case E: return N;
-      case SE: return NE;
-      case S: return E;
-      case SW: return SE;
-      case W: return S;
-      case NW: return SW;
+      case none: return none;
+      case n: return w;
+      case ne: return nw;
+      case e: return n;
+      case se: return ne;
+      case s: return e;
+      case sw: return se;
+      case w: return s;
+      case nw: return sw;
     }
 
     throw "unreachable";
@@ -75,15 +75,15 @@ class Direction extends VecBase implements Vec {
 
   Direction get rotateRight90 {
     switch (this) {
-      case NONE: return NONE;
-      case N: return E;
-      case NE: return SE;
-      case E: return S;
-      case SE: return SW;
-      case S: return W;
-      case SW: return NW;
-      case W: return N;
-      case NW: return NE;
+      case none: return none;
+      case n: return e;
+      case ne: return se;
+      case e: return s;
+      case se: return sw;
+      case s: return w;
+      case sw: return nw;
+      case w: return n;
+      case nw: return ne;
     }
 
     throw "unreachable";
@@ -91,15 +91,15 @@ class Direction extends VecBase implements Vec {
 
   Direction get rotate180 {
     switch (this) {
-      case NONE: return NONE;
-      case N: return S;
-      case NE: return SW;
-      case E: return W;
-      case SE: return NW;
-      case S: return N;
-      case SW: return NE;
-      case W: return E;
-      case NW: return SE;
+      case none: return none;
+      case n: return s;
+      case ne: return sw;
+      case e: return w;
+      case se: return nw;
+      case s: return n;
+      case sw: return ne;
+      case w: return e;
+      case nw: return se;
     }
 
     throw "unreachable";
@@ -107,15 +107,15 @@ class Direction extends VecBase implements Vec {
 
   String toString() {
     switch (this) {
-      case NONE: return "NONE";
-      case N: return "N";
-      case NE: return "NE";
-      case E: return "E";
-      case SE: return "SE";
-      case S: return "S";
-      case SW: return "SW";
-      case W: return "W";
-      case NW: return "NW";
+      case none: return "none";
+      case n: return "n";
+      case ne: return "ne";
+      case e: return "e";
+      case se: return "se";
+      case s: return "s";
+      case sw: return "sw";
+      case w: return "w";
+      case nw: return "nw";
     }
 
     throw "unreachable";