+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Make constants `lowerCamelCase`.
+* Remove `sign()` and `clamp()` since the Dart core libraries have them now.
+
 ## 0.1.1
 
 * `CARDINAL` and `INTERCARDINAL` direction lists to `Direction`.